# api/swagger.yaml
swagger: "2.0"
info:
  description: "This is the backend service for Bass OS. This service provides interfaces between the frontend and Hardware, OS and audio services."
  version: "1.0.0"
  title: "bassOS-service"
  license:
    name: "GPL v3.0"
    url: "https://www.gnu.org/licenses/gpl.html"
host: "10.10.10.1"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /settings/{setting-id}:
    get:
      tags:
      - "setting-status"
      summary: "Get the current status of a setting"
      operationId: "getSettingStatus"
      produces:
      - "application/json"
      parameters:
      - name: "setting-id"
        in: "path"
        description: "The id of the setting"
        required: true
        type: "integer"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SettingStatus"
        400:
          description: "invalid setting id"
      x-swagger-router-controller: "settings"
    put:
      tags:
      - "setting-status"
      summary: "Change the status of a setting"
      operationId: "setSettingStatus"
      consumes:
      - "application/json"
      parameters:
      - name: "setting-id"
        in: "path"
        description: "The id of the setting"
        required: true
        type: "integer"
      - in: "body"
        name: "setting-status"
        description: "The new status of the setting"
        required: true
        schema:
          $ref: "#/definitions/SettingStatus"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ServiceStatus"
        405:
          description: "validation Error"
        400:
          description: "invalid setting id"
      x-swagger-router-controller: "settings"
...
definitions:
  ServiceStatus:
    type: "object"
    properties:
      status:
        type: "string"
        pattern: "active|playing|stopped"
  SettingStatus:
    type: "object"
    properties:
      status:
        type: "boolean"
...